+++ packages/stream_core/lib/src/errors.dart
@@ -1,2 +1,3 @@
 export 'errors/stream_api_error.dart';
+export 'errors/stream_error_code.dart';
 export 'errors/stream_exception.dart';